UIA handler: add UIA drag drop effect property as a global property event. Re #12271. (#14278)
Link to issue number:
Follow-up to #12271
Follow-up to #14081

Summary of the issue:
Follow-up/quick fix engineering for drag and drop effect announcement: suppose a user moves the mouse to an object that: is a UIA element, supports drag and drop, and is not a system focus. When drag and drop operation starts, drag drop effect is not announced, caused by the drag drop effect not being a global property event.

Description of user facing changes
Drag drop effect will be announced when performing drag and drop operation via mouse regardless of system focus, similar to drop target effect announcement.

Description of development approach
Add UIA_DragDropEffectPropertyId to global properties list to align with drop target effect property.

Follow-up/quick fix engineering for drag and drop effect announcement: suppose a user moves the mouse to an object that: is a UIA element, supports drag and drop, and is not a system focus. When drag and drop operation starts, drag drop effect is not announced, caused by the drag drop effect not being a global property event.
Beta branch justification: this is a bug fix/new feature introduced in NVDA 2022.4. Subsequent testing indicate that drag drop UIA property should have been a global property so the property can be announced based on mouse actions, simiar to reporting mouse when mouse tracking is turned on via mouse settings panel. This change also aligns with drop target effect announcement as drag and drop announcement issue covers both.