Prototype for Windows Terminal: Use notifications instead of monitoring for new text (#14047)
Closes #13781

Summary of the issue:
Windows Terminal supports UIA notification events to notify of new text. This is way quicker than the current diffing we use to calculate new text to speak in a terminal.

Description of user facing changes
Note: This is hidden behind a feature flag.
When enabled, this should improve responsivity whereas having minimal user impact. The only thing I noticed is that when typing echo is on, the last typed character/word is spoken after pressing enter, whereas in the old situation, the last typed char/word was always silenced. I'd say this is an improvement.

Description of development approach
Handle the Windows Terminal object as regular Editable Text, i.e. remove the terminal specific stuff from it.