Handle loss of foreground during systemtests
If the chrome / notepad title can not be reported with NVDA+T during
the system tests, then use the task list/switcher to refocus the
target application.

Ensure that 'chrome' / 'notepad' is included in speech for the
application title.
Use regex to simplify matching aspects of title speech
that is required and ignore aspects that are irrelevant.

Note that the way that 'chrome' is presented may differ:
E.G:
* "- Chrome"
* "- Chrome Canary"
* "Chrome Canary" (without the dash)