Prevent double beep or 'cap' announcement with delayed character descriptions (#14271)
Fixes #14239

Summary of the issue:
There are 3 settings for notifying a user of a capital letter. There is "beep", announce 'cap', and/or shift pitch for capitals.

When navigating by character, there is already a beep or "cap" announcement.
When the delayed character description occurs, another beep or "cap" announcement occurs.
There is no need for a secondary beep or "cap" announcement when announcing the the delayed character description.

The cap pitch change may be useful, as a pitch change may be harder to notice, and continuing the shifted pitch is more intuitive.

Description of user facing changes
When beep or announce 'cap' for capitals is enabled with delayed character descriptions, the descriptions do not beep or announce 'cap' twice.

Description of development approach
Disable beeping and announcing 'cap' for delayed character descriptions.