UIA focus event handler: make sure to return if an NVDAObject could not be created. (#14314)
Fixes regression introduced in pr #14256

Summary of the issue:
Sometimes a UIA focus event handler can't create an NVDAObject for the element being focused. Normally, in this situation, the event would be silently dropped.
However, with the introduction of extra debugging for UIA events in pr #14256, a return was accidently removed, which meant that the shouldAllowUIAFocusEvent attribute tried to be looed up on the None (failed NVDAObject). This caused a noisy traceback in the log at level error.

Description of user facing changes
A noisy error traceback is no longer produced when NVDA cannot handle certain broken UIA focus events.

Description of development approach
Make sure to return if instantiating the NVDAObject produces None.