Support building liblouis on Clang 15 and above (#14319)
liblouis/liblouis#1259

Summary of the issue:
Starting with Clang 15, liblouis no longer compiles because Clang 15 treats int to pointer conversion as erroneous.

Description of user facing changes
NVDA builds properly again on VS 2022 preview.

Description of development approach
Added a flag that degrades the error to a warning again.

Testing strategy:
Ensure that NVDA builds correctly.

Known issues with pull request:
This is not strictly necessary at this moment because Clang 15 hasn't yet landed in stable. Please note however that we have no control over the appveyor release process and that building will instantly break without this patch as soon as Appveyor updates VS to a version that has Clang 15.