Make automatic braille display detection more pythonic (#14147)
Related to #14130

Summary of the issue:
Parts of the braille display detection code are hard to read and rely on Win32 APCs to be queued. A major drawback about APCS is that, as soon as theyre queued, they can not be canceled any more.

Description of user facing changes
None, though I hope things will be more stable after this change

Description of development approach
In an attempt to make the code more pythonic as well as changing the code in such a way that it no longer depends on the braille background thread, I changed the following:

The detection work now has a dedicated thread, using a thread pool under the hood. This is a thread pool with only one thread. This might sound stupid, but the major advantage of a ThreadPoolExecutor is that it allows to execute and queue functions on the thread
The detection function is no longer a loop. It was a loop to avoid queuing an apc when we knew one was already running, but as the ThreadPoolExecutor has an internal queue and we can cancel futures now, this decreases compexity.