Streamline importing of addon directories for python packages (#14350)
Closes #14340

Summary of the issue:
For addons, the several directories are added to the python package path using a function in the config module. This needs refactoring:

nvda/source/config/__init__.py, Line 487 in f674505
 # FIXME: this should not be coupled to the config module.... 

Description of user facing changes
None

Description of development approach
This pr changes logic to add the several addon directories to the package path in the addonHandler.