Map the selectable state for Java Access Bridge (#14351)
Fixes #14336

Summary of the issue:
In Java applications NVDA always announces selected for items which are selected. This can be quite annoying in certain controls such as menus. In these cases it is preferred to announce when something is not selected. The issue is that NVDA was not mapping the selectable state for Java Access Bridge. With this fix Java controls which have the selectable state will now only announce when something is not selected.

Description of user facing changes
For selectable controls in Java applications using the Java Access Bridge the user will now notice NVDA does not announce if something is selected, rather it will only announce when something is not selected.

Description of development approach
After discussing how to reduce the selected announcement verbosity in the associated issue, I decided to check what states were being exposed in other applications where it would announce items as "not selected". When reviewing the logs it appeared that the common factor is these controls had the selectable state. I am aware that Java Access Bridge exposes a selectable state. Upon inspecting the NVDA code for mapping Java Access Bridge objects to NVDA objects, I could not find any mapping for the selectable state. So I added the mapping and tested whether this fixed the issue.