No longer try to return to first task switch item (PR #14343)
Fixes: https://github.com/nvaccess/nvda/issues/14342

Summary of the issue:
When using the task switcher to navigate to the SUT application window, it is unnecessary to return to the first item in the task switcher. It is unlikely that there will be more than the current limit of 10 windows open on appveyor. If the test needs to return to the first window it will cycle through and return to the first row and column.

Development approach:
- Adds more debug logging to the robot framework log.
- Always cycle forwards through the task switcher.
- All failure paths send escape to dismiss the task switcher.
- Consider variations in task switcher content:
  * A single item. Unlikely during the tests, there should be the SUT application and the appveyor build agent.
  * A single row of items, most likely. Row is only reported when the row number changes. "column 1" indicates cycling back to the first item.
  * Multiple rows of items, unlikely on build systems, more likely on dev systems.