Symbol preserve property is honoured in the case of repeated symbols. (#14392)
Fixes #14391

Summary of the issue:
Symbol 'preserve' property was not honoured when the symbol is repeated more than 3 times. E.g. on the following line containing 4 dollar sign and at symbol level None:
$$$$

The synth does not read anything whereas it usually reads something when passed the dollar sign.

Description of user facing changes
When repeated, a symbol for which reporting level is below the current level is now passed to the synth is 'preserve' property is set to norep or always.