Removed unuseful empty or English-only documentation locale files (#14376)
* Removed unuseful documentation files for the following languages:
- Kanada (kn) change log and user guide
- Punjabi (pa) change log and user guide
- Hebrew (he) change log only
- Irish (ga) change log only
- Icelandic (is) user guide only

This files are either almost empty or contain only English content. Removing them allows to users of these languages to open up-to-date English documentation instead of empty or outdated English documentation.

Summary of the issue:
Some locale documentation files (change log or user guide) are empty or contain only out-of-date English text:

Since there is no content written in locale language in these files, the users of these languages have an empty documentation or an out-of-date English documentation.

If there is not translation work on these files for now (and there has actually never been except for Icelandic), it would be better for the users to have at least the up-to-date English documentation.

Description of user facing changes
Users will have up-to-date English documentation files instead of empty or out-of-date English-only documentation files.

Description of development approach
Remove English-only or empty files in the locale doc folders.