Do not ignore locale CLDR file when locale symbol file is missing (#14433)
In NVDA some languages do not have a symbol.dic file but have a cldr.dic file. Today, these are the following languages: 'af_ZA', 'am', 'as', 'gu', 'id', 'kok', 'ml', 'mni', 'ne', 'te', 'th', 'ur'.
For these languages the symbols belonging to CLDR file (e.g. "☺") are reported in English rather than in the locale language.
This is because we detect if a language has no symbol file and in this case, we store it in _noSymbolLocalesCache and raise a LookupError in the future, even if the language has a CLDR file.

Description of user facing changes
Languages with no symbol file will be able to report CLDR in the locale language if available.

Description of development approach
Use two caches, one for languages with no symbol file and one for languages with no CLDR file.