Follow up for decide_enabled extension point (#14536)
Fixup of #14503

Summary of the issue:
When a handler was registered to decide_enabled and a cursor was blinking or a message with timeout was shown on the braille display, the display would still be updated. Found when working on NVDA Remote to integrate the extension points.
The extensionPointTestHelpers wouldn't strictly test whether the handler was actually called. Also the handler argument for value was incorrect and using a class to store the filter value was overdone.
Description of user facing changes
When a handler decides to disable the braille handler, cursor blinking is now stopped and a message is dismissed.