Avoid a situation where no NVDA modifier key is defined (#14528)
Link to issue number:
Fixes #14527

May be considered a follow-up of #14233.

Summary of the issue:
When configuring NVDA Modifier Key for a non-default profile and for the default profile, we can end-up with no NVDA key defined at all in the non-default profile (see #14527 for detailed steps).

This is due to the fact that the config stores one item per NVDA key but that there is a dependency between these values to honor the requirement that at least one key should be defined as NVDA key. These values are checked when changing NVDA keys via the GUI. But the requirement may be missed when switching profiles and stacking their configurations.

There is also a second scenario allowing to have no NVDA key defined using only default profile:

In keyboard settings enable only caps lock and disable both insert, then validate
Open the NVDA startup dialog via Help menu, uncheck caps lock and validate.
This scenario is quite unlikely but there is no check in NVDA startup dialog to prevent such situation.

Description of user facing changes
Situations where no NVDA key is defined will not be possible anymore.

Description of development approach
Use only one config item to store the used NVDA modifier keys.
Define a config flag for NVDA key; each value corresponds to one bit so that all the values can be bitwise or-ed and the reesulting value can be stored in the config as an integer.