appModules/Notepad: reposition status bar index from 7 to -2. Re #14573.
Upcoming Notepad version 11.2212 introduces tabbed interface, useful for opening multiple files in a single Notepad window. As a result, this changes the UI tree in that status bar entry position is no longer the eighth position. Thankfully, the status bar text is the second to last item in the Notepad UI tree across versions (stable and dev), so use -2 to represent status bar index.