Old scripts remained in gestures.ini will not prevent anymore new gestures to be displayed in the INput gesture dialog (#14600)
Fixes #14599

Summary of the issue:
In some case (e.g. NVDA dev or add-on dev/update/uninstall), the file gesture.ini may contain script association for scripts that do not exist anymore in the current config. As described in #14599, if you try to reassign the shortcut that was used for such old removed scripts, the gesture may not be displayed in the input gesture dialog.

Description of user facing changes
When the file gesture.ini contains script association for unexisting scripts, e.g. due to old add-on or old devs in NVDA source code, there will not be anymore the risk that a gesture associated to a currently valid script be not displayed.

Description of development approach
• If a script name is not found in a class, do not consider this script handled and remove it from the handledScripts set; also log a debugWarning in this case.
• And also checked other continue statement (and commented one of them) to check the logic with respect to scripts added in this list.