Avoid circular reference on IOBase (#14605)
It is possible to register a callback on hwIo.IoBase to catch read errors. However, if this is a callback on a driver and the driver closes the device, the device keeps a reference to the callback on the driver. This causes a circular reference.

Description of user facing changes
Possibly less unreachable object errors

Description of development approach
Similar to _onReceive, ensure that onReadError is set to None when closing the device.