Allow translators to specify row/column when reporting the end coordinate of a merged cell (#14585)
For some translators (Chinese ones at least), in the case of merged cell it would be better to report "column 2 to column 3" instead of "column 2 to 3" as we do today.
Today, NVDA uses 2 translatable strings to report the coordinates of a merged cell:
• for rows: "row %s" and "through %s"
• for columns: "column %s" and and "through %s"

Description of user facing changes
Translators will now have two strings indicating the end coordinate of a merged cell, one for rows and one for columns.

Description of development approach
I just have used .format(...) with named parameters. Defining strings with two parameters named differently allows to create 2 translatable strings instead of only one. As a side effect, it's also better for translators to have named parameters rather than just %s which is not explicit.
Note: using pgettext was another option, but with named parameters, it didn't turn out to be necessary.