Refactor BrailleHandler.setDisplayByName (#14524)
Summary of the issue:
braille.BrailleHandler.setDisplayname is too complex. It also relies on reentrancy and is difficult to grasp.

Description of user facing changes
See bug fixes below

Description of development approach
setDisplayByName is now split in three functions
setDisplayByName will no longer set noBraille as a fallback if noBraille is already active. This avoids reinitializing noBraille multiple times
BrailleHandler._detectionEnabled is removed, as it was supposed to be equal to the boolean value of BrailleHandler._detector and that value was checked anyway. Note that that's supposed to be private API
When configuration was set to auto detect, BrailleHandler.handleDisplayUnavailable relied on behavior in setDisplayByName to re-enable the detector when falling back to no braille. This logic is now simplified.