Fix regexp error message in dictionaries (#14557)
Fixes #14556

Summary of the issue:
Dictionaries do not allow to use backslash character normally in the replacement field when regexp is not selected as entry type. When such a character is used, a message box complaining of a regexp error appears.

Description of user facing changes
The user can use backslash in replacement field normally when entry type is not regexp
Error messages are more precise and more specific when entering erroneous values in the fields for a regexp entry
Description of development approach
When the entry type is not regexp, escape the replacement string
When entries are created:
filter exception on re.error instead of the too global Exception to catch only exceptions related to regexps
separate the try/except statement in two in order to indicate in which field the regexp error lies
In case of a regexp error, after validating "OK" in the message box, the field where the error lies is re-focused
Reports the regexp error only in case of entry type = regexp; in the other cases,re-raise the exception, in order not to catch unrelated exceptions, e.g. code bug.