Browse mode: only filter out speaking of focus events caused by NVDA, rather than trying to work out if the caret is still within the last focus. (#14611)
NVDA sometimes fails to report focus in browseMode documents, as it is trying to filter out redundant focus events caused by moving the caret.
Particular cases:
• Focus moves from a control inside a gridcell or interactive list item, back to the ancestor gridcell or interactive list item.
• Focus moves from a gridcell or interactive list item, to the first control inside the gridcell or interactive list item.
Currently, NVDA chooses to ignore any focus in browseMode that overlaps with the browseMode caret. This is so that any focus event that results from moving the caret is not reported as this would be redundant. However, this rule is clearly too limiting as it stops the above examples from working.

Description of user facing changes
In Browse mode, NVDA will no longer incorrectly ignore focus moving to a parent or child control e.g. moving from a control to its parent lit item or gridcell.
Note however that this fix only applies when the Automatically set focus to focusable elements option in Browse Mode settings is turned off (which is the default).

Description of development approach
Rather than ignoring focus changes in browse mode if the new focus's range overlaps the existing caret position, focus is only ignored if it is the same object that NvDA has specifically requested setting focus to, such as when pressing the applications key or activating a control.