Use Word object model to report distances when available (#14575)
When pressing NVDA+F, for Chinese people, paragraph indent is reported in number of characters by default when NVDA is using Word object model to access Word documents. But it is reported in centimeter when NVDA is using UIA to access Word documents.
When using object model, the way NVDA reports indent in Word honours the two following advanced options of Word:
• "Show measurements in" (centimeters, inches, etc.)
• "Show measurements in character width" (Note: for this option to be available, you need to have an East-Asian language installed in Word's languages)
These options should be also honoured when using UIA for Word documents. Especially, Chinese people are used to express measurements in characters, what NVDA does not when using UIA for Word documents.

Description of user facing changes
Word's advanced options are now used instead of system options also when using UIA to access Word documents. Thus, the option "Show measurements in character width" in Word advanced options is now honoured also when using UIA for Word documents.

Description of development approach
The only way I've found to know how distances are reported in Word (according to this option) is to use Word object model. Thus, if possible, we use Word object model to access Word's options, even if NVDA uses UIA to access Word document.
When object model is not available however, Windows regional setting is used as a fallback as it was before this PR.