Gecko vbuf: Don't render anything if the document is dead.
NVDA kills buffers for dead documents when a focus event is received.
However, sometimes, when the user closes a tab, the closed document might die after the focus event for the newly active tab.
Also, the focus event can be delayed sometimes; e.g. if loading a new document takes a while.
When this happens, the buffer is still alive and listening for events, but the ids of nodes in that buffer might be reused by other documents!
Trying to render those (completely arbitrary) nodes could cause a broken buffer tree, leading to instability later.

To fix this, hold a COM reference to the document.
When an event is received which ostensibly matches a node in the buffer, check if the document is alive.
If it isn't, skip the node and clear the buffer.
To avoid excessive COM calls, only check this once per buffer update tick.