Fix-up of #14557: remove unwanted backslashes appearing in replacement expression of dictionary (#14664)
Fixes #14663
Fix-up of #14557

Summary of the issue:
When reading text containing a non-regexp dictionary entry, if the replacement field contains some special characters (such as space, +, etc.), these characters are escaped with a backslash. Thus this backslash is heard in the text spoken by NVDA.

Description of user facing changes
No unappropriated backslash will be spoken by NVDA when the replacement field of a dictionary entry contains space or other special character.

Description of development approach
Use theString.replace('\\', '\\\\') instead of re.escape(theString).

Indeed, it was well documented in the doc of re.escape:

This function must not be used for the replacement string in sub() and subn(), only backslashes should be escaped.