Import keyLabels only after translation is  initialized to make the labels displayed in the  user's language (#14658)
Fixes #14657

Summary of the issue:
In pull request #14528 keyLabels were imported pretty early during NVDA's startup. Since this module contains localized strings at the module level these are translated before languageHandler is initialized. This means that they're translated either to the default language of the system, or for locales for which Python's locale.getdefaultlocale fails they remain in English, disregarding the language set in preferences.

Description of user facing changes
Key labels are once again presented in the language set in preferences.

Description of development approach
keyLabels are imported lazily in the configFlags
I've added docstring to the keyLabels module explaining when it can be safely imported.
Testing strategy:
On a English Windows set NVDA's language to Polish, ensured that key labels in the keyboard help and in the preferences are displayed in Polish.