Changed regex split string from ";" to "__NVDA:split-here__", in ui.browseableMessage, for less ambiguity when separating title from message. (#14668)
Fixes #14667

Summary of the issue:
If ui.browseableMessage was given a title containing a semicolon (;), the title would be broken at the semicolon, and the rest of the title would be rendered before (as part of) the message in the HTML portion of the window.
This had the potential to result in unexpected outcomes, if the message content was HTML; not to mention being confusing for the user.

While this was noticed in relation to the Report link destination feature introduced by #14583, it did not result from that feature, and could have occurred prior to its introduction. It does however effect it significantly, as the title of the link the destination of which is being identified, is presented in the title of a ui.browseableMessage, if the command is called twice, or the alternate version of the command is mapped to a key.

Description of user facing changes
If a link name contains a semicolon, portions of that name will no longer appear along with the link address, when reporting link destination in its window version.

Description of development approach
Changed the regex separator string in ui.browseableMessage, and in the message.html file, to one which should hopefully be much less likely to appear in the wild. Instead of a single semicolon (";"), the string is now "__NVDA:split-here__".

Also removed the blank line from the top of message.html. I know of no purpose for having a blank line there.